# openapi.yaml
openapi: 3.1.0
info:
  description: LensBridge backend — media upload and MusallahBoard management
  title: LensBridge API
  version: 0.0.1-SNAPSHOT
servers:
- url: /
security:
- bearerAuth: []
paths:
...
  /api/admin/audit/upload/{uploadId}:
    get:
      operationId: getUploadAuditHistory
      parameters:
      - in: path
        name: uploadId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditEventDto"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      tags:
      - admin-controller
...
  /api/admin/board/events/{eventId}:
    delete:
      operationId: deleteEvent
      parameters:
      - in: path
        name: eventId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      tags:
      - board-admin-controller
    get:
      operationId: getBoardEventById
      parameters:
      - in: path
        name: eventId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardEvent"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      summary: Fetch a single board event
      tags:
      - board-admin-controller
    patch:
      operationId: updateEvent
      parameters:
      - in: path
        name: eventId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCalendarEventRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BoardEvent"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      tags:
      - board-admin-controller
...
  /api/admin/upload/{uploadId}:
    delete:
      operationId: deleteUpload
      parameters:
      - in: path
        name: uploadId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      tags:
      - admin-controller
    post:
      operationId: approveUpload
      parameters:
      - in: path
        name: uploadId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      tags:
      - admin-controller
...
  /api/events:
    get:
      operationId: getPublicEvents
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MediaEvent"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      summary: List publicly visible events
      tags:
      - events-controller
  /api/events/{id}:
    get:
      operationId: getPublicEventById
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MediaEvent"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      summary: Fetch a single publicly visible event
      tags:
      - events-controller
  /api/gallery:
    get:
      operationId: getGalleryUploads
      parameters:
      - description: Zero-based page index (0..N)
        in: query
        name: page
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - description: The size of the page to be returned
        in: query
        name: size
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
      - description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        in: query
        name: sort
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageGalleryItemDto"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      summary: Page through approved gallery media
      tags:
      - gallery-controller
  /api/gallery/event/{eventId}:
    get:
      operationId: getGalleryByEvent
      parameters:
      - in: path
        name: eventId
        required: true
        schema:
          type: string
          format: uuid
      - description: Zero-based page index (0..N)
        in: query
        name: page
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - description: The size of the page to be returned
        in: query
        name: size
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
      - description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        in: query
        name: sort
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageGalleryItemDto"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      tags:
      - gallery-controller
...
  /api/upload/event/{eventId}:
    get:
      operationId: getUploadsByEvent
      parameters:
      - in: path
        name: eventId
        required: true
        schema:
          type: string
          format: uuid
      - description: Zero-based page index (0..N)
        in: query
        name: page
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - description: The size of the page to be returned
        in: query
        name: size
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
      - description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        in: query
        name: sort
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageUploadDto"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      tags:
      - file-upload-controller
  /api/upload/limits:
    get:
      operationId: getUploadLimits
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadLimitsResponse"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      tags:
      - file-upload-controller
  /api/upload/{eventId}/direct/complete:
    post:
      operationId: completeDirectUpload
      parameters:
      - in: path
        name: eventId
        required: true
        schema:
          type: string
          format: uuid
      - in: query
        name: objectKey
        required: true
        schema:
          type: string
      - in: query
        name: filename
        required: true
        schema:
          type: string
      - in: query
        name: contentType
        required: true
        schema:
          type: string
      - in: query
        name: fileSize
        required: true
        schema:
          type: integer
          format: int64
      - in: query
        name: instagramHandle
        required: false
        schema:
          type: string
      - in: query
        name: description
        required: false
        schema:
          type: string
      - in: query
        name: anon
        required: false
        schema:
          type: boolean
          default: false
      - in: query
        name: expectedSha256
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadCompletionResponse"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      tags:
      - file-upload-controller
  /api/upload/{eventId}/direct/presign:
    post:
      operationId: generatePresignedUploadUrl
      parameters:
      - in: path
        name: eventId
        required: true
        schema:
          type: string
          format: uuid
      - in: query
        name: filename
        required: true
        schema:
          type: string
      - in: query
        name: contentType
        required: true
        schema:
          type: string
      - in: query
        name: fileSize
        required: true
        schema:
          type: integer
          format: int64
      - in: query
        name: expectedSha256
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PresignedUploadResponse"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      tags:
      - file-upload-controller
  /api/upload/{uploadId}:
    get:
      operationId: getUploadById
      parameters:
      - in: path
        name: uploadId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadDto"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      tags:
      - file-upload-controller
...
  /api/user/uploads/{uploadId}:
    delete:
      operationId: deleteUserUpload
      parameters:
      - in: path
        name: uploadId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      tags:
      - user-controller
components:
  schemas:
...
    AuditEventDto:
      type: object
      properties:
        action:
          type: string
          enum:
          - APPROVE_UPLOAD
          - UNAPPROVE_UPLOAD
          - DELETE_UPLOAD
          - FEATURE_UPLOAD
          - UNFEATURE_UPLOAD
          - CREATE_EVENT
          - UPDATE_EVENT
          - DELETE_EVENT
          - CREATE_CALENDAR_EVENT
          - UPDATE_CALENDAR_EVENT
          - DELETE_CALENDAR_EVENT
          - CREATE_POSTER
          - UPDATE_POSTER
          - DELETE_POSTER
          - CREATE_SOCIAL
          - UPDATE_SOCIAL
          - DELETE_SOCIAL
          - SAVE_WEEKLY_CONTENT
          - DELETE_WEEKLY_CONTENT
          - UPDATE_BOARD_CONFIG
          - UPDATE_BOARD_TICKER
          - REFRESH_BOARDS
          - ISSUE_ENROLLMENT_TOKEN
          - REVOKE_DEVICE
          - ISSUE_DEVICE_COMMAND
          - PROMOTE_USER
          - DEMOTE_USER
          - DISABLE_USER
          - ENABLE_USER
          - VIEW_AUDIT_LOGS
          - EXPORT_DATA
          - SYSTEM_MAINTENANCE
          - VERIFY_USER
          - UNVERIFY_USER
          - RESET_USER_PASSWORD
          - TRIGGER_PASSWORD_RESET_EMAIL
          - ADD_USER_ROLE
          - REMOVE_USER_ROLE
          - GRANT_PERMISSION
          - REVOKE_PERMISSION
          - ADD_USER
          - REMOVE_USER
          - UPDATE_USER
        adminEmail:
          type: string
        adminId:
          type: string
          format: uuid
        adminName:
          type: string
        details:
          type: string
        id:
          type: string
          format: uuid
        ipAddress:
          type: string
        targetEntityId:
          type: string
          format: uuid
        targetEntityType:
          type: string
          enum:
          - USER
          - UPLOAD
          - MUSALLAH_BOARD
          - DEVICE
          - EVENT
        timestamp:
          type: string
          format: date-time
        userAgent:
          type: string
    BoardEvent:
      type: object
      properties:
        allDay:
          type: boolean
        audience:
          type: string
          enum:
          - brothers
          - sisters
          - both
        description:
          type: string
        endTime:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        location:
          type: string
        name:
          type: string
        startTime:
          type: string
          format: date-time
...
    GalleryItemDto:
      type: object
      properties:
        author:
          type: string
        date:
          type: string
        event:
          type: string
        featured:
          type: boolean
        id:
          type: string
        src:
          type: string
        thumbnail:
          type: string
        title:
          type: string
        type:
          type: string
...
    MediaEvent:
      type: object
      properties:
        date:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum:
          - UPCOMING
          - ONGOING
          - PAST
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    PageGalleryItemDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/GalleryItemDto"
        empty:
          type: boolean
        first:
          type: boolean
        last:
          type: boolean
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/Pageable"
        size:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
    PageUploadDto:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/UploadDto"
        empty:
          type: boolean
        first:
          type: boolean
        last:
          type: boolean
        number:
          type: integer
          format: int32
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: "#/components/schemas/Pageable"
        size:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
...
    PresignedUploadResponse:
      type: object
      properties:
        contentType:
          type: string
        eventId:
          type: string
          format: uuid
        expectedSha256:
          type: string
        expiresInMinutes:
          type: integer
          format: int32
        method:
          type: string
        objectKey:
          type: string
        uploadUrl:
          type: string
...
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
...
    UpdateCalendarEventRequest:
      type: object
      properties:
        allDay:
          type: boolean
        audience:
          type: string
          enum:
          - brothers
          - sisters
          - both
        description:
          type: string
        endTime:
          type: string
          format: date-time
        location:
          type: string
        name:
          type: string
        startTime:
          type: string
          format: date-time
...
    UploadCompletionResponse:
      type: object
      properties:
        eventId:
          type: string
          format: uuid
        fileSize:
          type: integer
          format: int64
        objectKey:
          type: string
        uploadId:
          type: string
          format: uuid
        verified:
          type: boolean
    UploadDto:
      type: object
      properties:
        anon:
          type: boolean
        approved:
          type: boolean
        contentType:
          type: string
          enum:
          - image
          - video
          - audio
          - document
        createdDate:
          type: string
          format: date-time
        eventId:
          type: string
          format: uuid
        eventName:
          type: string
        featured:
          type: boolean
        fileName:
          type: string
        fileUrl:
          type: string
        instagramHandle:
          type: string
        thumbnailUrl:
          type: string
        uploadDescription:
          type: string
        uploadedBy:
          type: string
          format: uuid
        uuid:
          type: string
          format: uuid
    UploadLimitsResponse:
      type: object
      properties:
        allowedContentTypes:
          type: array
          items:
            type: string
        dailyLimit:
          type: integer
          format: int32
        maxSizeBytes:
          type: integer
          format: int64
        maxSizeMB:
          type: integer
          format: int64
        role:
          type: string
        uploadsRemaining:
          type: integer
          format: int64
        uploadsToday:
          type: integer
          format: int64
        videoMaxDurationSeconds:
          type: integer
          format: int32
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      name: bearerAuth
      scheme: bearer
      type: http